% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_strings.R
\name{clean_strings}
\alias{clean_strings}
\title{Clean string variables in ROAR data}
\usage{
clean_strings(df, verbose = TRUE)
}
\arguments{
\item{df}{The dataframe that needs to be cleaned}

\item{verbose}{A Boolean value, TRUE (default) or FALSE, that controls whether
informational messages and warnings are displayed}
}
\value{
A dataframe with NAs rather than empty strings and no extra characters in assigning organization variables
}
\description{
Remove extra characters from assigning organization variables and convert
empty strings to NA values.
}
\examples{
test_df <- data.frame(assigning_schools = c("[irNgj3c]", "irNgj3c"),
                      age = c("", "6.7"))
clean_df <- clean_strings(test_df)
}
