% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_grade.R
\name{estimate_grade}
\alias{estimate_grade}
\title{Estimate grade using age in months}
\usage{
estimate_grade(age_months_at_run)
}
\arguments{
\item{age_months_at_run}{The variable that is associated with age in months}
}
\value{
A dataframe with estimated and corrected grades
}
\description{
If grade is missing or incorrect, use this function to estimate grade based
on age in months at the time of the run.
}
\examples{
\dontrun{
# estimate grade for anyone before 24-25 school year because of incorrect grades
df <- df \%>\% mutate(user_grade = dplyr::case_when(time_started < as.Date("2024-07-31") ~
                                            map_chr(age_months, estimate_grade), TRUE ~ user_grade))
# estimate grade for anyone with missing grade
df <- df \%>\% mutate(user_grade = dplyr::case_when(is.na(user_grade) ~
                    purrr::map_chr(age_months, estimate_grade), TRUE ~ user_grade))
}
}
