% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_assessments.R
\name{filter_assessments}
\alias{filter_assessments}
\title{Filter completed, best, and reliable runs}
\usage{
filter_assessments(
  df,
  completed = TRUE,
  reliable = FALSE,
  best_run = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{df}{The dataframe that contains all runs including incomplete, non-best,
and unreliable runs}

\item{completed}{A Boolean value, TRUE (default) or FALSE, that indicates if
the researcher wants to remove incomplete runs}

\item{reliable}{A Boolean value, TRUE or FALSE (default), that indicates if
the researcher wants to remove unreliable runs. Note: this feature is only
implemented in the ROAR core assessments}

\item{best_run}{A Boolean value, TRUE or FALSE (default), that indicates if
the researchers wants to keep only the best run within each administration.
Note: this feature is only implemented in some of the ROAR assessments}

\item{verbose}{A Boolean value, TRUE (default) or FALSE, that controls whether
informational messages and warnings are displayed}
}
\value{
A dataframe that has properly filtered the indicated runs
}
\description{
Filters for completed, best, and reliable runs based on selection. Completed runs
are always set to TRUE (default) because this feature is implemented in most
(if not all) ROAR assessments. Best and reliables runs are set to FALSE (default)
as these are only implemented in the ROAR core assessments.
}
\examples{
test_df <- data.frame(
           task_id = c("roam-alpaca", "swr", "sre", "letter", "sre-es", "swr-es"),
           completed = c("true", "true", "false", "true", "false", "false"),
           best_run = c(NA, "true", "false", "true", NA, NA),
           reliable = c(NA, "true", "false", "true", NA, NA))
clean_df <- filter_assessments(test_df)
clean_df <- filter_assessments(test_df, completed=TRUE, best_run=TRUE, reliable=TRUE)
}
