% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_histogram.R
\name{plot_scatter_histogram}
\alias{plot_scatter_histogram}
\title{Create scatter plot with marginal histograms}
\usage{
plot_scatter_histogram(
  df,
  grade = "user_grade_at_run",
  prop_correct = "prop_correct",
  median_rt = "median_rt",
  verbose = TRUE
)
}
\arguments{
\item{df}{The dataframe that contains grade, proportion correct,
and median response time.}

\item{grade}{A string value to indicate the name of the variable representing
the student's grade (default: "user_grade_at_run")}

\item{prop_correct}{A string value to indicate the name of the variable
representing the proportion correct for each student's run (default:
"prop_correct")}

\item{median_rt}{A string value to indicate the name of the variable
representing the median response time the student's run (default:
"median_rt")}

\item{verbose}{A Boolean value, TRUE (default) or FALSE, that controls whether
informational messages and warnings are displayed}
}
\value{
A marginal distribution plot showing the relationship between a ROAR
assessment proportion correct and the median response time
}
\description{
Uses grade, proportion correct, and median response time to create a
marginal distribution plot. The main plot is a scatter plot showing
the relationship between proportion correct and median response time.
The marginal histograms are bar charts along the top and right edges of the
main plot that show the distribution of each variable independently.
}
