% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Remove duplicate rows}
\usage{
remove_duplicates(df)
}
\arguments{
\item{df}{The dataframe that has identical rows}
}
\value{
A dataframe that contains only unique rows
}
\description{
Removing duplicate rows ensures researchers are accurately counting each unique
student, run, or trial.
}
