% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roar.read.csv.R
\name{roar.read.csv}
\alias{roar.read.csv}
\title{Read ROAR data and remove opt-outs}
\usage{
roar.read.csv(data_name, data_path, opt_out_link, verbose = TRUE)
}
\arguments{
\item{data_name}{Name of the data file (string)}

\item{data_path}{Path to the directory where data file exists (string)}

\item{opt_out_link}{Google Drive link to most up-to-date opt-out file (string)}

\item{verbose}{A Boolean value, TRUE (default) or FALSE, that controls whether
informational messages and warnings are displayed}
}
\value{
A dataframe with opt-outs removed
}
\description{
Read ROAR data using read.csv() and remove opt-outs from the df using the
provided Google Drive link. Filters the opt-outs in "PID", "pid",
"assessment_pid", "user.assessmentPid" columns.
}
\examples{
\dontrun{
new_data <- roar.read.csv("all_runs.csv", "~/Documents",
"google.drive.link")


# Silent mode
new_data <- roar.read.csv("all_runs.csv", "~/Documents",
"google.drive.link", verbose = FALSE)
}
}
