% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_grade.R
\name{standardize_grade}
\alias{standardize_grade}
\title{Standardize grade to uniform values}
\usage{
standardize_grade(df, grade_col = "user_grade_at_run")
}
\arguments{
\item{df}{The dataframe that has inconsistent grade values}

\item{grade_col}{The grade variable that needs to be standardized
(e.g., "user_grade_at_run" (default))}
}
\value{
A dataframe with uniform grade values
}
\description{
Create uniform values for grade (e.g., "2", "two", and "2nd" all
become "2").
}
