% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_grade.R
\name{estimate_grade}
\alias{estimate_grade}
\title{Estimate grade using age in months}
\usage{
estimate_grade(age_months_at_run, verbose = TRUE)
}
\arguments{
\item{age_months_at_run}{The variable that is associated with age in months}

\item{verbose}{A Boolean value, TRUE (default) or FALSE, that controls whether
informational messages and warnings are displayed}
}
\value{
A dataframe with estimated and corrected grades
}
\description{
If grade is missing or needs to be corrected, use this function to estimate grade based
on age in months at the time of the run.
}
\examples{
\dontrun{
library(tidyverse)
test_df <- data.frame(
  age_months = c(75, 83, 99, 200),
  user_grade = c(NA, "2", "2", "10"),
  time_started = c("2025-03-12", "2024-02-17", "2026-01-09", "2025-04-19"))
clean_df <- test_df \%>\% dplyr::mutate(user_grade = dplyr::case_when(
  time_started < as.Date("2024-07-31") ~ purrr::map_chr(age_months, estimate_grade),
  TRUE ~ user_grade))
clean_df <- clean_df \%>\% dplyr::mutate(user_grade = case_when(
  is.na(user_grade) ~ purrr::map_chr(age_months, estimate_grade),
  TRUE ~ user_grade))
}
}
