% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_accounts.R
\name{remove_accounts}
\alias{remove_accounts}
\title{Remove test, demo, pilot, QA, or NA accounts}
\usage{
remove_accounts(
  df,
  test = TRUE,
  demo = TRUE,
  pilot = TRUE,
  qa = TRUE,
  na = FALSE
)
}
\arguments{
\item{df}{The dataframe that contains all types of accounts}

\item{test}{A Boolean value, TRUE (default) or FALSE, that indicates whether
or not to remove test accounts/participants}

\item{demo}{A Boolean value, TRUE (default) or FALSE, that indicates whether
or not to remove demo accounts/participants}

\item{pilot}{A Boolean value, TRUE (default) or FALSE, that indicates whether
or not to remove pilot accounts/participants}

\item{qa}{A Boolean value, TRUE (default) or FALSE, that indicates whether
or not to remove QA accounts/participants}

\item{na}{A Boolean value, TRUE or FALSE (default), that indicates whether
or not to remove NA assessment_pid/accounts/participants}
}
\value{
A dataframe with the indicated remaining account types
}
\description{
Remove test, demo, pilot, QA, or NA accounts
}
