% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_empty_cols.R
\name{remove_empty_cols}
\alias{remove_empty_cols}
\title{Remove columns with all NA values}
\usage{
remove_empty_cols(df, verbose = TRUE)
}
\arguments{
\item{df}{The dataframe that contains extra columns containing only NA values}

\item{verbose}{A Boolean value, TRUE (default) or FALSE, that controls whether
informational messages and warnings are displayed}
}
\value{
A dataframe with columns that contain some non-NA values
}
\description{
There are many columns within the ROAR dataframes that only apply to specific assessments.
Removing the columns that only contain NA values, allows researchers to isolate
important variables.
}
\examples{
test_df <- data.frame(firstname = c("Jane", "John", NA, "Kelly"),
                      lastname = c("Doe", NA, NA, "Smith"),
                      middlename = c(NA, NA, NA, NA))
clean_df <- remove_empty_cols(test_df)
}
