% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roar.read.csv.R
\name{roar.read.csv}
\alias{roar.read.csv}
\title{Read ROAR data and remove opt-outs}
\usage{
roar.read.csv(data_name, data_path, opt_out_link)
}
\arguments{
\item{data_name}{Name of the data file}

\item{data_path}{Path to the data file}

\item{opt_out_link}{Google Drive link to most up-to-date opt-out file}
}
\value{
A data frame with opt-outs removed
}
\description{
Read ROAR data and remove opt-outs
}
