% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roar.read.csv.R
\name{roar.read.csv}
\alias{roar.read.csv}
\title{Read ROAR data and remove opt-outs}
\usage{
roar.read.csv(data_name, data_path, opt_out_link)
}
\arguments{
\item{data_name}{Name of the data file}

\item{data_path}{Path to the data file}

\item{opt_out_link}{Google Drive link to most up-to-date opt-out file}
}
\value{
A dataframe with opt-outs removed
}
\description{
Read ROAR data using read.csv() and remove opt-outs from the df using the
provided Google Drive link. Filters the opt-outs in "PID", "pid",
"assessment_pid", "user.assessmentPid" columns.
}
\examples{
\dontrun{
new_data <- roar.read.csv("all_runs.csv", "~/Documents",
"https://drive.google.com/file/d/11gYLqU5xT-NMDxWXGQj8WfZ8AVA_lFT9/view?usp=drive_link")
}
}
